+++ static/tpl/current/frailty.js
@@ -1,11 +1,20 @@
 'use strict';
 
-var modified = false;
+var Utils = new Utils(),
+	modified = false,
+	datas = {};
 
 window.addEventListener('DOMContentLoaded', function() {
+	moment.locale(Cookies.get("lang")=="en"?"en-gb":Cookies.get("lang"));
 	document.addEventListener('change', function( evt ) {
 		modified = true;
 	}, false );
+
+	//init lockdown
+    datas.validateStatus = (document.querySelector('#validate-status').innerHTML === 'true');
+    if(datas.validateStatus) {
+    	Utils.lockdown();
+    }
 	
 }, false);
 
@@ -25,21 +34,18 @@ function checkForm(validate) {
 	formObj.risk = (formObj.choice === 'risk');
 
 	formObj.validated = validate;
+	if(validate) {
+		formObj.validatedDate = moment().format('YYYY-MM-DD');
+	}
 
 	delete formObj.choice;
 
 	promiseXHR('PUT', '../api/beneficiary/current/frailty', 200, JSON.stringify(formObj))
 		.then(function(res) {
-			if (JSON.parse(res).validated) {
-				document.getElementById('buttons').innerHTML = '';
-
-				var inputs = document.querySelectorAll('input');
-				for (var i = 0; i < inputs.length; ++i) {
-					inputs[i].setAttribute('disabled', true);
-				}
-			}
-			new Modal('saveSuccess');
-			modified = false;
+			new Modal('saveSuccess', function() {
+				modified = false;
+				window.location.href = '/current/frailty';
+			});
 		}, function(error) {
 			new Modal('errorOccured');
 			console.log(error);